+++ src/components/Popup/GetHelp/GetHelp.module.scss
@@ -10,11 +10,11 @@
 }
 
 .form {
-  max-width: 500px;
   display: flex;
   flex-direction: column;
   justify-content: center;
   align-items: center;
+  max-width: 500px;
   text-align: center;
   color: $color-darkGrey;
